consolidate amenities
